+++ rust/app/Cargo.toml
@@ -5,7 +5,7 @@ edition = "2021"
 
 [dependencies]
 reim_wrap = { path = "../reim_wrap" }
-reim_dt_wasm = { path = "../reim_dt_wasm" }
+reim_host_dt_wasm = { path = "../reim_host_dt_wasm" }
 reim_codegen = { path = "../reim_codegen" }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0.85"
