+++ rust/app/src/wrappers/external_references/Cargo.toml
@@ -7,8 +7,8 @@ license = "MIT"
 edition = "2021"
 
 [dependencies]
-reim_wrap = { path = "../../../../reim_wrap" }
-reim_wrap_wasm = { path = "../../../../reim_wrap_wasm" }
+reim_wrapper_dt_wasm = { path = "../../../../reim_wrapper_dt_wasm" }
+reim_dt = { path = "../../../../reim_dt" }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0.85"
 async-trait = "0.1.58"
