+++ rust/app/src/wrappers/external_references/src/polywrap/external/classes/test_external_class.rs
@@ -1,6 +1,6 @@
 use std::sync::Arc;
 
-use reim_wrap::{ ExternalModule };
+use reim_dt::ExternalModule;
 use serde::{Serialize, Deserialize};
 use serde_json::json;
 use crate::polywrap::external::module::wrap_module::get_external_module_or_panic;
