+++ rust/app/src/wrappers/external_references/src/polywrap/external/module/internal_wrap_module.rs
@@ -1,7 +1,7 @@
 use std::sync::Arc;
 
 use async_trait::async_trait;
-use reim_wrap::{ ExternalModule, InternalModule };
+use reim_dt::{ ExternalModule, InternalModule };
 use crate::polywrap::internal::{ invoke_global_function, invoke_class_method };
 use crate::polywrap::resources::InternalResource;
