+++ rust/reim_host_dt_wasm/src/utils.rs
