+++ rust/reim_host_dt_wasm/src/wasm_runtime/imports.rs
@@ -1,5 +1,6 @@
-use std::sync::{Arc, Mutex};
+use std::sync::Arc;
 
+use futures::lock::Mutex;
 use wasmtime::{AsContextMut, Caller, Linker, Memory};
 
 use crate::{error::WrapperError, wasm_runtime::instance::State};
@@ -12,22 +13,24 @@ fn write_to_memory(buffer: &mut [u8], ptr: usize, data: &[u8]) {
     buffer[ptr..(ptr + data.len())].copy_from_slice(data);
 }
 
-pub fn create_imports(
+pub async fn create_imports(
     linker: &mut Linker<State>,
     arc_memory: &Arc<Mutex<Memory>>,
 ) -> Result<(), WrapperError> {
     let memory = Arc::clone(&arc_memory);
     linker
-        .func_wrap(
+        .func_wrap1_async(
             "wrap",
             "__dt_fill_input_buffer",
             move |mut caller: Caller<'_, State>, buffer_ptr: u32| {
-                let memory = memory.lock().unwrap();
-                let (memory_buffer, state) = memory.data_and_store_mut(caller.as_context_mut());
 
-                write_to_memory(memory_buffer, buffer_ptr as usize, &state.input_buffer);
-
-                Ok(())
+                let memory = Arc::clone(&memory);
+                Box::new(async move {
+                    let memory = memory.lock().await;
+                    let (memory_buffer, state) = memory.data_and_store_mut(caller.as_context_mut());
+    
+                    write_to_memory(memory_buffer, buffer_ptr as usize, &state.input_buffer);
+                })
             },
         )
         .map_err(|e| WrapperError::WasmRuntimeError(e.to_string()))?;
@@ -40,11 +43,10 @@ pub fn create_imports(
             move |mut caller: Caller<'_, State>,
                   buffer_ptr: u32,
                   buffer_len: u32| {
-                let memory = memory.lock().unwrap();
-                let async_memory = Arc::new(tokio::sync::Mutex::new(*memory));
 
+                let memory = Arc::clone(&memory);
                 Box::new(async move {
-                    let memory = async_memory.lock().await;
+                    let memory = memory.lock().await;
                     let (memory_buffer, state) = memory.data_and_store_mut(caller.as_context_mut());
                     let buffer =
                         read_from_memory(memory_buffer, buffer_ptr as usize, buffer_len as usize);
@@ -69,23 +71,24 @@ pub fn create_imports(
 
     let memory = Arc::clone(&arc_memory);
     linker
-        .func_wrap(
+        .func_wrap1_async(
             "wrap",
             "__dt_fill_send_result",
             move |mut caller: Caller<'_, State>, buffer_ptr: u32| {
-                let memory = memory.lock().unwrap();
-                let (memory_buffer, state) = memory.data_and_store_mut(caller.as_context_mut());
+                let memory = Arc::clone(&memory);
+                Box::new(async move {
+                    let memory = memory.lock().await;
+                    let (memory_buffer, state) = memory.data_and_store_mut(caller.as_context_mut());
 
-                write_to_memory(memory_buffer, buffer_ptr as usize, &state.send_result);
-
-                Ok(())
+                    write_to_memory(memory_buffer, buffer_ptr as usize, &state.send_result);
+                })
             },
         )
         .map_err(|e| WrapperError::WasmRuntimeError(e.to_string()))?;
 
     let memory = Arc::clone(&arc_memory);
     linker
-        .define("env", "memory", *memory.lock().unwrap())
+        .define("env", "memory", *memory.lock().await)
         .map_err(|e| WrapperError::WasmRuntimeError(e.to_string()))?;
 
     Ok(())
