+++ rust/reim_wrap/Cargo.toml
@@ -5,6 +5,5 @@ edition = "2021"
 
 [dependencies]
 async-trait = "0.1.58"
-tokio = { version = "1", features = ["full"] }
-reim_dt = { path = "../reim_dt" }
 futures = "0.3.25"
+reim_dt = { path = "../reim_dt" }
