--- rust/reim_wrap/src/lib.rs
@@ -1,7 +1,0 @@
-#![feature(async_closure)]
-
-pub mod internal_module;
-pub mod external_module;
-pub use internal_module::*;
-pub use external_module::*;
-pub mod wrapper_module;
