+++ rust/reim_wrapper_dt_wasm/Cargo.toml
@@ -1,11 +1,9 @@
 [package]
-name = "reim_dt_wasm"
+name = "reim_wrapper_dt_wasm"
 version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-wasmtime = "1.0.1"
+futures = "0.3.25"
 reim_dt = { path = "../reim_dt" }
-thiserror = "1.0.37"
 async-trait = "0.1.58"
-tokio = { version = "1", features = ["full"] }
