+++ rust/reim_wrapper_dt_wasm/src/lib.rs
@@ -3,6 +3,6 @@ pub mod receive;
 pub mod imports;
 pub mod malloc;
 pub mod abort;
-pub mod host_wrap_module;
+pub mod host_module;
 pub use receive::*;
-pub use host_wrap_module::*;
+pub use host_module::*;
