+++ rust/reim_wrapper_dt_wasm/src/malloc.rs
